api/pyproject.toml
[tool.poetry]
name = "nrg-flight-api"
version = "0.1.0"
description = "FastAPI backend for a simple airplane ticketing service"
authors = ["NRG Flight <dev@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = { version = "^0.30.0", extras = ["standard"] }

# DB (async)
SQLAlchemy = { version = "^2.0.36", extras = ["asyncio"] }
asyncpg = "^0.30.0"

# Migrations
alembic = "^1.14.0"

# Config
pydantic-settings = "^2.6.1"

# Auth / Security
python-jose = { version = "^3.3.0", extras = ["cryptography"] }
passlib = { version = "^1.7.4", extras = ["bcrypt"] }
email-validator = "^2.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
httpx = "^0.27.2"
aiosqlite = "^0.20.0"
ruff = "^0.6.9"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-q --cov=app --cov-report=term-missing --cov-fail-under=100"

[tool.coverage.run]
branch = false
source = ["app"]
omit = [
  "app/alembic/*",
  "app/alembic/versions/*",
]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 100

[tool.ruff]
line-length = 100
